requirements.txt
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
sqlalchemy==1.4.41
prefect==2.10.8
pytest==7.3.1
requests==2.28.2
python-dotenv==0.21.1
loguru==0.7.0 